+++ src/syntax/block/olist.test.ts
@@ -14,6 +14,7 @@ describe('Unit: syntax/olist', () => {
       assert.deepStrictEqual(inspect(parser('0.\n*')), void 0);
       assert.deepStrictEqual(inspect(parser('0.\n0')), void 0);
       assert.deepStrictEqual(inspect(parser('0.\n 0.\n0')), void 0);
+      assert.deepStrictEqual(inspect(parser('0.\n  0.\n 0.')), void 0);
       assert.deepStrictEqual(inspect(parser(' 0.')), void 0);
     });
 
@@ -39,6 +40,7 @@ describe('Unit: syntax/olist', () => {
       const parser = loop(olist);
       assert.deepStrictEqual(inspect(parser('0. 1\n 0. 2')), [['<ol start="1"><li>1<ol start="1"><li>2</li></ol></li></ol>'], '']);
       assert.deepStrictEqual(inspect(parser('0. 1\n 0. 2\n 0. 3')), [['<ol start="1"><li>1<ol start="1"><li>2</li><li>3</li></ol></li></ol>'], '']);
+      assert.deepStrictEqual(inspect(parser('0. 1\n 0. 2\n  0. 3')), [['<ol start="1"><li>1<ol start="1"><li>2<ol start="1"><li>3</li></ol></li></ol></li></ol>'], '']);
       assert.deepStrictEqual(inspect(parser('0. 1\n 0. 2\n0. 3')), [['<ol start="1"><li>1<ol start="1"><li>2</li></ol></li><li>3</li></ol>'], '']);
       assert.deepStrictEqual(inspect(parser('0. *1*')), [['<ol start="1"><li><em>1</em></li></ol>'], '']);
     });