+++ src/syntax/block/ulist.test.ts
@@ -14,6 +14,7 @@ describe('Unit: syntax/ulist', () => {
       assert.deepStrictEqual(inspect(parser('-\n*')), void 0);
       assert.deepStrictEqual(inspect(parser('-\n0')), void 0);
       assert.deepStrictEqual(inspect(parser('-\n -\n0')), void 0);
+      assert.deepStrictEqual(inspect(parser('-\n  -\n -')), void 0);
       assert.deepStrictEqual(inspect(parser(' -')), void 0);
     });
 
@@ -39,6 +40,7 @@ describe('Unit: syntax/ulist', () => {
       const parser = loop(ulist);
       assert.deepStrictEqual(inspect(parser('- 1\n - 2')), [['<ul><li>1<ul><li>2</li></ul></li></ul>'], '']);
       assert.deepStrictEqual(inspect(parser('- 1\n - 2\n - 3')), [['<ul><li>1<ul><li>2</li><li>3</li></ul></li></ul>'], '']);
+      assert.deepStrictEqual(inspect(parser('- 1\n - 2\n  - 3')), [['<ul><li>1<ul><li>2<ul><li>3</li></ul></li></ul></li></ul>'], '']);
       assert.deepStrictEqual(inspect(parser('- 1\n - 2\n- 3')), [['<ul><li>1<ul><li>2</li></ul></li><li>3</li></ul>'], '']);
       assert.deepStrictEqual(inspect(parser('- *1*')), [['<ul><li><em>1</em></li></ul>'], '']);
     });